pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "imoia"
version = "0.2.0"
description = "Plataforma de gestão de investimento imobiliário fix and flip (evolução do ImoScout)"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "sqlalchemy>=2.0.0",
    "streamlit>=1.40.0",
    "plotly>=5.24.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "tenacity>=9.0.0",
    "zeep>=4.2.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "celery[redis]>=5.4",
    "pydantic-settings>=2.0",
    "psycopg2-binary>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
